🧹 [Code Health] Refactor draw_diamond to use bresenham algorithm
