--- commit_msg.txt
@@ -1,16 +1,0 @@
-⚡ Optimize `export_trimmed` allocations and handle multi-byte slices properly
-
-💡 **What:**
-- Refactored `export_trimmed` to pre-calculate capacity and iterate grid directly into a single string.
-- Limited the column loop iteration using `max_width`.
-- Used character iteration safely instead of relying on byte slicing (`&line[..options.max_width]`) which panicked on multi-byte emoji boundaries.
-
-🎯 **Why:**
-- The previous implementation allocated an entire string for the grid, collected its `.lines()` into a `Vec<&str>`, and then iterated again to apply `max_width` creating a final limited String. This caused O(N) intermediate allocations.
-- A user encountering emojis or non-ASCII chars could trigger a fatal panic during byte slicing.
-
-📊 **Measured Improvement:**
-- Benchmark runtime on 200 iterations over a 500x500 grid:
-  - Baseline: ~1.70s
-  - Optimized (single-pass): ~1.03s
-  - Performance improved by ~40% while fixing multi-byte edge cases and minimizing memory allocations.
