+++ package.json
@@ -35,7 +35,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
-    "@playwright/test": "^1.58.2",
+    "@playwright/test": "^1.59.1",
     "eslint": "^10.3.0",
     "playwright": "^1.58.2",
     "typescript-eslint": "^8.59.2"
